edugenie/requirements.txt
# requirements.txt
streamlit
fastapi
pydantic
psycopg2-binary
faiss-cpu
sentence-transformers
PyYAML
fitz  # PyMuPDF
nltk
requests
python-dotenv